+++ src/shared/api/file/index.ts
@@ -1,0 +1,7 @@
+import { DeleteFile } from '@/shared/api/file/type';
+import axiosInstance from '@/shared/api/instance';
+
+export const deleteTimeBlock = async (data: DeleteFile) => {
+  const response = await axiosInstance.post('/file', data);
+  return response;
+};
