+++ src/shared/api/file/type.ts
@@ -1,0 +1,3 @@
+export interface DeleteFile {
+  fileName: string;
+}
