+++ src/shared/api/file/upload/index.ts
@@ -1,0 +1,10 @@
+import axiosInstance from '@/shared/api/instance';
+
+export const getFile = async (fileFormat: string) => {
+  const response = await axiosInstance.get('/file/upload', {
+    params: {
+      fileFormat: fileFormat,
+    },
+  });
+  return response.data.data;
+};
