+++ src/shared/api/time-blocks/team/time-block/index.ts
@@ -1,0 +1,11 @@
+import axiosInstance from '@/shared/api/instance';
+import { TimeBlockCreate } from '@/shared/api/time-blocks/team/time-block/type';
+
+export const postTimeBlock = async (teamId: number, type: string, data: TimeBlockCreate) => {
+  const response = await axiosInstance.post(`/time-blocks/team/${teamId}/time-block`, data, {
+    params: {
+      type: type,
+    },
+  });
+  return response.data;
+};
