+++ src/shared/api/time-blocks/team/time-block/type.ts
@@ -1,0 +1,8 @@
+export interface TimeBlockCreate {
+  name: string;
+  color: string;
+  startDate: string;
+  endDate: string;
+  blockType: string;
+  files: Map<string, string>;
+}
